+++ xarray/test/test_backends.py
@@ -986,10 +986,11 @@ class H5NetCDFDataTest(BaseNetCDF4Test, TestCase):
     def test_encoding_unlimited_dims(self):
         ds = Dataset({'x': ('y', np.arange(10.0))})
         ds.encoding = {'unlimited_dims': ['y']}
-        with pytest.warns(UserWarning):
-            ds.to_netcdf('foo-bar.nc', engine='h5netcdf')
-        with pytest.warns(UserWarning):
-            ds.to_netcdf('foo-bar.nc', engine='h5netcdf', unlimited_dims=['y'])
+        with create_tmp_file() as tmp_file:
+            with pytest.warns(UserWarning):
+                ds.to_netcdf(tmp_file, engine='h5netcdf')
+            with pytest.warns(UserWarning):
+                ds.to_netcdf(tmp_file, engine='h5netcdf', unlimited_dims=['y'])
 
 
 @requires_dask